# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RaktOra REST API",
    "description": "Comprehensive REST API specification for RaktOra - National Voluntary Blood Network and Smart FEFO Inventory Engine.",
    "version": "1.0.0",
    "contact": {
      "name": "RaktOra Developer Team",
      "email": "support@raktora.org"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token"
      }
    },
    "schemas": {
...
      "BloodRequestInput": {
        "type": "object",
        "required": ["patient_name", "blood_group", "units_needed", "hospital_name", "city", "state", "contact_number"],
        "properties": {
          "patient_name": { "type": "string", "example": "Suresh Verma" },
          "blood_group": { "type": "string", "enum": ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"], "example": "B+" },
          "component": { "type": "string", "enum": ["WHOLE_BLOOD", "PRBC", "PLATELETS", "FFP", "CRYOPRECIPITATE", "SDP"], "example": "WHOLE_BLOOD" },
          "units_needed": { "type": "integer", "minimum": 1, "example": 2 },
          "hospital_name": { "type": "string", "example": "Apollo Hospital" },
          "hospital_address": { "type": "string", "example": "Bannerghatta Road" },
          "city": { "type": "string", "example": "Bengaluru" },
          "state": { "type": "string", "example": "Karnataka" },
          "urgency_level": { "type": "string", "enum": ["Normal", "Urgent", "Critical"], "example": "Urgent" },
          "contact_number": { "type": "string", "example": "9876543210" }
        }
      },
      "FulfillRequestInput": {
        "type": "object",
        "required": ["units"],
        "properties": {
          "units": { "type": "integer", "minimum": 1, "example": 1 }
        }
      },
...
    }
  },
  "paths": {
...
    "/requests": {
      "get": {
        "summary": "List Emergency Blood Requests",
        "tags": ["Blood Requests"],
        "responses": {
          "200": { "description": "Requests list" }
        }
      },
      "post": {
        "summary": "Create Emergency Blood Request",
        "tags": ["Blood Requests"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BloodRequestInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Request created successfully" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/requests/{id}/fulfill": {
      "post": {
        "summary": "Fulfill Blood Request Units (Atomic)",
        "tags": ["Blood Requests"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FulfillRequestInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Fulfillment contribution applied" },
          "400": { "description": "Invalid quantity or request already fulfilled" },
          "404": { "description": "Request not found" }
        }
      }
    },
...
  }
}